Split the last three oversized screens and delete maxLinesAllowlist (#28)
RaceSetup 487→254, Progress 394→232, RaceResults 348→212. Both allowlists in
eslint.config.mjs are now gone, and the 300-line cap applies to every view
module with no exceptions.

The seams are the panels, because that's what these screens already are — a
column of independently comprehensible sections:

  setup/PresetRow      the one-tap answers
  setup/DeckPicker     which numbers land on the cards, or the drill that
                       replaces them
  setup/ClockPicker    time per card, presets plus the exact field
  setup/RivalList      who you're racing, and Start
  results/Scoreline    the five numbers you look at first
  results/Rewards      level, bonuses, badges — renders nothing when there's
                       nothing to celebrate
  results/SplitsTable  card by card against the rival
  progress/FactMap     144 facts by mastery
  progress/RecordBook  best times, mine against the household's, plus RunList

Two things nearly went wrong in the move and are worth recording. `factKey`,
`masteryOf` and `Mastery` live in engine/records, not engine/progress, which
type-checking caught. And I retyped the mastery legend as "Not tried /
Learning / Solid / Mastered" — the real copy is "Not tried yet / Still
learning / Getting there / Mastered". Nothing would have failed; the wording
is just softer, deliberately, for a child reading that their times tables are
"Not tried yet" rather than "Not tried". There's now an assertion on it.

Verified in a browser, because a section that silently stops rendering doesn't
error:

  Progress   all six panels present · 144 fact cells · legend wording exact
             · records and run list populated · switching operation keeps the map
  Results    five stats · 10 split rows · timed-out rows marked · 5 actions
  Setup      the clock field still snaps and clamps (3.3→3.25, 99→60, 0.1→1)
  plus the full smoke run, all with zero console errors

Closes #4
