+++ src/components/screens/Progress.tsx
@@ -2,11 +2,10 @@ import { useMemo, useState } from "react";
 import { Navigate, useNavigate, useParams } from "react-router-dom";
 import { useHub, usePlayer } from "@/components/state/HubContext";
 import TopBar from "@/components/TopBar";
-import { Avatar, Button } from "@/components/ui/kit";
+import { Button } from "@/components/ui/kit";
 import { BADGES } from "@/engine/progress";
 import {
   bestRun,
-  factKey,
   factStats,
   lifetimeStats,
   masteryOf,
@@ -14,27 +13,17 @@ import {
   sessionsFor,
   tableProgress,
   troubleFacts,
-  type Mastery,
 } from "@/engine/records";
 import {
   OPERATIONS,
-  OPERATION_ORDER,
   buildDrill,
-  describeConfig,
   timeLimitForAge,
 } from "@/engine/decks/flashcards";
 import { sfx } from "@/services/sound";
-import { clock, duration, percent, shortDate } from "@/engine/format";
+import { FactMap } from "./progress/FactMap";
+import { RecordBook, RunList } from "./progress/RecordBook";
+import { duration, percent } from "@/engine/format";
 import type { Operation } from "@/engine/types";
-
-const AXIS = Array.from({ length: 12 }, (_, i) => i + 1);
-
-const MASTERY_LABEL: Record<Mastery, string> = {
-  untried: "Not tried yet",
-  learning: "Still learning",
-  solid: "Getting there",
-  mastered: "Mastered",
-};
 
 export default function Progress() {
   const { profileId } = useParams();
@@ -115,90 +104,12 @@ export default function Progress() {
         </div>
       </section>
 
-      <section className="panel anim-rise">
-        <div className="panel__head">
-          <h2 className="panel__title">Fact map</h2>
-          <div className="segmented segmented--slim">
-            {OPERATION_ORDER.map((op) => (
-              <Button
-                key={op}
-                variant="bare"
-                className={`segmented__btn${operation === op ? " is-on" : ""}`}
-                onClick={() => setOperation(op)}
-                pressed={operation === op}
-              >
-                {OPERATIONS[op].symbol}
-              </Button>
-            ))}
-          </div>
-        </div>
-        <p className="muted">
-          Every {spec.label.toLowerCase()} fact from 1 to 12. A square turns
-          green once it&apos;s answered right three times running and under four
-          seconds — that&apos;s recall, not counting.
-        </p>
-
-        <div className="factmap__wrap">
-          <table className="factmap">
-            <caption className="u-sr">{spec.label} facts by mastery</caption>
-            <thead>
-              <tr>
-                <th scope="col" className="factmap__corner u-mono">
-                  {spec.symbol}
-                </th>
-                {AXIS.map((n) => (
-                  <th key={n} scope="col" className="u-mono factmap__axis">
-                    {n}
-                  </th>
-                ))}
-              </tr>
-            </thead>
-            <tbody>
-              {AXIS.map((row) => (
-                <tr key={row}>
-                  <th scope="row" className="u-mono factmap__axis">
-                    {row}
-                  </th>
-                  {AXIS.map((col) => {
-                    const stat = grid.get(factKey(row, col));
-                    const level = masteryOf(stat);
-                    const avg = stat
-                      ? stat.totalMs / stat.attempts / 1000
-                      : null;
-                    return (
-                      <td key={col} className={`factmap__cell is-${level}`}>
-                        <span className="u-sr">
-                          {row} {spec.symbol} {col}: {MASTERY_LABEL[level]}
-                        </span>
-                        <span className="factmap__tip" aria-hidden="true">
-                          {row} {spec.symbol} {col}
-                          {stat
-                            ? ` · ${stat.correct}/${stat.attempts} · ${avg!.toFixed(1)}s`
-                            : " · not tried"}
-                        </span>
-                      </td>
-                    );
-                  })}
-                </tr>
-              ))}
-            </tbody>
-          </table>
-        </div>
-
-        <ul className="legend">
-          {(["untried", "learning", "solid", "mastered"] as Mastery[]).map(
-            (level) => (
-              <li key={level} className="legend__item">
-                <span
-                  className={`legend__swatch is-${level}`}
-                  aria-hidden="true"
-                />
-                {MASTERY_LABEL[level]}
-              </li>
-            ),
-          )}
-        </ul>
-      </section>
+      <FactMap
+        operation={operation}
+        onOperationChange={setOperation}
+        spec={spec}
+        grid={grid}
+      />
 
       <section className="panel anim-rise">
         <div className="panel__head">
@@ -328,86 +239,9 @@ export default function Progress() {
         </section>
       </div>
 
-      <section className="panel anim-rise">
-        <div className="panel__head">
-          <h2 className="panel__title">Records</h2>
-        </div>
-        {records.length === 0 ? (
-          <p className="muted">Race something and your best times land here.</p>
-        ) : (
-          <div className="splits__scroll">
-            <table className="splits">
-              <thead>
-                <tr>
-                  <th scope="col">Race</th>
-                  <th scope="col">Your best</th>
-                  <th scope="col">House best</th>
-                </tr>
-              </thead>
-              <tbody>
-                {records.map((record) => {
-                  const isMine = record.houseBest.profileId === profile.id;
-                  return (
-                    <tr key={record.key}>
-                      <td className="splits__card">
-                        {describeConfig(record.myBest.config)}
-                      </td>
-                      <td className="u-mono">
-                        {clock(raceTimeMs(record.myBest))}
-                      </td>
-                      <td className="u-mono splits__holder">
-                        {clock(raceTimeMs(record.houseBest))}
-                        {record.holder && (
-                          <span
-                            className={`splits__who${isMine ? " is-ahead" : ""}`}
-                          >
-                            <Avatar profile={record.holder} size="1.1rem" />
-                            {isMine ? "you" : record.holder.name}
-                          </span>
-                        )}
-                      </td>
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
-          </div>
-        )}
-      </section>
+      <RecordBook records={records} profileId={profile.id} />
 
-      <section className="panel anim-rise">
-        <div className="panel__head">
-          <h2 className="panel__title">Every race</h2>
-        </div>
-        {mine.length === 0 ? (
-          <p className="muted">No races yet.</p>
-        ) : (
-          <ul className="runlist">
-            {[...mine]
-              .reverse()
-              .slice(0, 40)
-              .map((session) => (
-                <li key={session.id} className="runlist__row">
-                  <span className="runlist__when u-mono">
-                    {shortDate(session.finishedAt)}
-                  </span>
-                  <span className="runlist__what">
-                    {describeConfig(session.config)}
-                  </span>
-                  <span className="runlist__acc u-mono">
-                    {percent(
-                      session.correct /
-                        Math.max(1, session.correct + session.incorrect),
-                    )}
-                  </span>
-                  <span className="runlist__time u-mono">
-                    {clock(raceTimeMs(session))}
-                  </span>
-                </li>
-              ))}
-          </ul>
-        )}
-      </section>
+      <RunList sessions={mine} />
     </main>
   );
 }
